requirements.txt
python-dotenv

# telegram
python-telegram-bot

#ukassa
yookassa
